refactor(test): dedupe script-test harness - AbstractQuarantineScriptTest
The duplicate-code check correctly flagged this PR's two script tests
sharing 24 lines of process-runner scaffolding (ProcessBuilder + readFully
+ Result). Extracted to AbstractQuarantineScriptTest: fixture root, registry
writer, runScript() with a customizeEnvironment() hook (the owners test uses
it for its PATH-stubbed gh/git). Script tests now contain only fixtures and
assertions. 31/31 self-tests green.
