+++ parallel-consumer-core/src/test/java/io/confluent/parallelconsumer/CheckQuarantineOwnersScriptTest.java
@@ -5,18 +5,13 @@ package io.confluent.parallelconsumer;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.condition.DisabledOnOs;
-import org.junit.jupiter.api.condition.OS;
-import org.junit.jupiter.api.io.TempDir;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.PosixFilePermissions;
-import java.util.concurrent.TimeUnit;
+import java.util.Map;
 
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.Truth.assertWithMessage;
@@ -28,11 +23,8 @@ import static com.google.common.truth.Truth.assertWithMessage;
  * closed-unmerged / merged-but-still-quarantined are ERRORS; transient gh failure, unowned entries,
  * base-not-updated, preview-keeps-annotation, and file-missing-from-preview are ADVISORY only.
  */
-@DisabledOnOs(OS.WINDOWS) // bash script under test; CI and dev machines are POSIX
-class CheckQuarantineOwnersScriptTest {
+class CheckQuarantineOwnersScriptTest extends AbstractQuarantineScriptTest {
 
-    @TempDir
-    Path fixture;
     private Path stubBin;
 
     @BeforeEach
@@ -150,53 +142,19 @@ class CheckQuarantineOwnersScriptTest {
 
     // ---- helpers ----
 
-    private void writeRegistry(String entries) throws IOException {
-        Path docs = fixture.resolve("docs");
-        Files.createDirectories(docs);
-        String content = "# registry\n\n## Currently quarantined\n\n" + entries;
-        Files.write(docs.resolve("QUARANTINED_TESTS.md"), content.getBytes(StandardCharsets.UTF_8));
-    }
-
     private void writeStub(String name, String body) throws IOException {
         Path stub = stubBin.resolve(name);
         Files.write(stub, body.getBytes(StandardCharsets.UTF_8));
         Files.setPosixFilePermissions(stub, PosixFilePermissions.fromString("rwxr-xr-x"));
     }
 
+    @Override
+    protected void customizeEnvironment(Map<String, String> env) {
+        env.put("PATH", stubBin + ":" + env.get("PATH"));
+        env.put("STUB_PHASE_FILE", fixture.resolve("stub-phase").toString());
+    }
+
     private Result run(String... envPairs) throws IOException, InterruptedException {
-        Path script = RepoRoot.find().resolve("bin/check-quarantine-owners.sh");
-        ProcessBuilder pb = new ProcessBuilder("bash", script.toString());
-        pb.environment().put("QUARANTINE_CHECK_ROOT", fixture.toString());
-        pb.environment().put("PATH", stubBin + ":" + pb.environment().get("PATH"));
-        pb.environment().put("STUB_PHASE_FILE", fixture.resolve("stub-phase").toString());
-        for (int i = 0; i < envPairs.length; i += 2) {
-            pb.environment().put(envPairs[i], envPairs[i + 1]);
-        }
-        pb.redirectErrorStream(true);
-        Process process = pb.start();
-        String output = readFully(process.getInputStream());
-        assertWithMessage("script hung - output so far: %s", output)
-                .that(process.waitFor(60, TimeUnit.SECONDS)).isTrue();
-        return new Result(process.exitValue(), output);
-    }
-
-    private static String readFully(InputStream in) throws IOException {
-        ByteArrayOutputStream buffer = new ByteArrayOutputStream();
-        byte[] chunk = new byte[4096];
-        int n;
-        while ((n = in.read(chunk)) != -1) {
-            buffer.write(chunk, 0, n);
-        }
-        return new String(buffer.toByteArray(), StandardCharsets.UTF_8);
-    }
-
-    private static final class Result {
-        final int exitCode;
-        final String output;
-
-        Result(int exitCode, String output) {
-            this.exitCode = exitCode;
-            this.output = output;
-        }
+        return runScript("bin/check-quarantine-owners.sh", envPairs);
     }
 }
